tela nova

+++ controleretorno/src/main/java/br/com/doctorfeet/controleretorno/model/Cliente.java
@@ -21,7 +21,7 @@ public class Cliente {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private Integer id;
+    private Long id;
 
     private String nome;
 
@@ -46,15 +46,15 @@ public class Cliente {
 
     public Cliente(){}
     
-    public Cliente(Integer id) {
+    public Cliente(Long id) {
         this.id = id;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
